requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
fastapi-cache2==0.2.1
